supervisor: fix supervisor startup when custom certs are involved
In this commit
https://github.com/balena-os/meta-balena/commit/b3222e619bbeebc0a2bd9e7419a9d8ae12b34dce
we removed `REGISTRY_ENDPOINT` from our sourced variables. Since this
value is still used to pass custom certs to the supervisor, we need to
keep it around for the time being. It can be removed once again as soon
as https://github.com/balena-os/meta-balena/issues/1398 is addressed.

Change-type: patch
Signed-off-by: Matthew McGinn <matthew@balena.io>