os-helpers-fs: Add functions to centralize access to fat files
The FAT filesystem is used for the boot partition to provide access from
all operating systems (Linux, Windows, Mac). This allows users to modify
its contents using the Balena CLI for example. It is also a simple enough
filesystem to be supported by all bootloaders, and mandatory for UEFI.

However, renames are not atomic on FAT filesystems unlike ext4, so if the
device loses power while a file is being renamed it could finish up in a
corrupted state.

This commit introduces a centralized safe_modify function that checksums
and backs up the file in question before the rename so it can be recovered
on boot if the power is lost.

Changelog-entry: Provide safe modification for files in FAT filesystems
Change-type: patch
Signed-off-by: Alex Gonzalez <alexg@balena.io>