fsuuidsinit: Wait for root device to come up before the initial check
If the rootfs is on a slow-to-bring-up device (e.g. RPi4 + USB)
the fsuuidsinit_enabled() function may return before the balena symlinks
are created. This gets wrongly interpreted as missing UUIDs leading to
a chain of failures in the subsequent scripts.
This patch moves the symlink waiting loop from fsuuidsinit_run()
to fsuuids_enabled() before the first usage of the symlink.

Changelog-Entry: Wait for the root device to come up when necessary
Change-Type: patch
Signed-off-by: Michal Toman <michalt@balena.io>