hostapp-update-hooks: Blacklist supervisor configuration backends
The hostapp update process should not overwrite the supervisor configuration
backend files to avoid the supervisor being forced to set the target state
after HUP and reboot the device during the rollback period.

This only applies to the host configuration files which are the only ones
that force a reboot.

Connects-to: https://github.com/balena-io/balena-supervisor/issues/1464
Change-type: patch
Changelog-entry: Blacklist supervisor configuration backend files during HUP
Signed-off-by: Alex Gonzalez <alexg@balena.io>