systemd/timeinit: add fake.hwclock to maintain system time over reboots
In order to produce sensible timestamps for journald log messages:
a) the system time needs to be maintained correctly over a reboot, and
b) the system time needs to be set before journald is started.

Currently the system time is maintained over reboots on systems
without an RTC using the last modified time of the chrony drift file.
However there are a couple of issues with this approach:

a) /var/lib/chrony/ is not mounted early enough in the boot process
to be available for setting the time before journald is started.
b) there is an issue with the current systemd dependencies that result
in the last modified time of the drift file not being updated when the
system is shutdown or rebooted (see #1995).

The Debian fake-hwclock service (as used by Raspberry Pi OS) has been
added to overcome these issues.

The fake-hwclock service will save and restore the system time from
the fake-hwclock.data file (in /etc/fake-hwclock/). The system time
is loaded from this file at boot and saved to it on shutdown. An
additional timer service has been added to update the file on an
hourly basis to cater for unexpected shutdown scenarios, e.g. power
failure.

The timeinit-timestamp service has improved logging and has been
renamed to timeinit-buildtime for clarity.

A new time-set.target has been added as per upstream systemd and
the fake-hwclock and timeinit-buildtime services have been added
to it.

Change-type: minor
Connects-to: #1367 #1919
Signed-off-by: Mark Corbin <mark@balena.io>