start-resin-supervisor: fix directory creation for 'balena start'
Ensure that the /var/log/supervisor-log directory exists prior to
running 'balena start --attach resin_supervisor' as well as
'runSupervisor'.

Updates to the 'start-resin-supervisor' script in v2.62.1 removed
the check for updates to the REGISTRY_ENDPOINT variable. Previously
this had been detected as changing every time the script was run
due to a comparison between 'null' and ''. This resulted in the
'start-resin-supervisor' script always running through the
'runSupervisor' path. With this variable check removed, and no config
updates being detected, the script was trying to run
'balena start --attach resin_supervisor' and failing due to the
absence of the /var/log/supervisor-log directory. To fix this problem we
unconditionally test for and create this directory (if necessary) so
that it is available regardless of the path taken through the script.

Change-type: patch
Connects-to: #2064
Changelog-entry: start-resin-supervisor: fix directory creation for 'balena start'
Signed-off-by: Mark Corbin <mark@balena.io>