start-resin-supervisor: populate NODE_EXTRA_CA_CERTS when custom CA is defined
This is necessary because node has its own built-in CA bundle and ignores
the system-wide bundle. Bind-mount the system-wide bundle to the supervisor
container as well to keep the previous behavior. Make it read-only though,
any modifications should be performed by the host OS.

Change-type: patch
Changelog-entry: Respect custom CA in supervisor
Signed-off-by: Michal Toman <michalt@balena.io>