networkmanager: Update to 1.28.0
Fixes #2075

Needed were a number of various changes to make the package compile properly:

Removed is 0001-wwan-Set-MTU-based-on-what-ModemManager-exposes.patch that is now
included upstream.

Our patch for removing HTTPS connectivity checking warning is reworked for ease of
maintainance. It now keeps the log entry, but changes it to debug level.

Fixed are UPSTREAM_CHECK_* definitions as they referenced a wrong version number.

The following additional configuration options were added/removed:

*. Introspection is disabled through `--enable-introspection=no`. Other services do
not depend on it, so it is safe to remove it. A related patch is no longer needed
0002-Do-not-create-settings-settings-property-documentati.patch

*. A new option for using firewalld zone for shared mode is disabled as we do not
use firewalld.

*. The polkit agent option no longer is available, so `--enable-polkit-agent` and
`--disable-polkit-agent` are no longer defined.

Change-type: minor
Changelog-entry: Update NetworkManager to 1.28.0
Signed-off-by: Zahari Petkov <zahari@balena.io>