ppp: Fix timeout when launched by NetworkManager
Fixes #1846

UART modems (RaspberryPi HATs) are not working currently under
balenaOS as NetworkManager times out while attempting to establish
ppp connection. This is not a balenaOS specific issue.

This commits adds a `connect ""` declaration to `/etc/ppp/options`
to workaround this as the NULL default value causes the timeout.

The connect option specifies an external script to establish the
physical link. When using NetworkManager/ModemManager it is
ModemManager that establishes the physical link before passing it
to NetworkManager. Thus `connect` should be empty.

Change-type: patch
Changelog-entry: Fix pppd timeout when launched by NetworkManager
Signed-off-by: Zahari Petkov <zahari@balena.io>