grub-efi: build required modules into grub image
grub-efi requires modules to be installed in the boot partition, and
resin-image installs them from ${DEPLOYDIR}.

A normal grub installation installs those modules to ${PREFIX}/${libdir}
to allow grub tooling to install them at runtime, but we're building the
image with GRUB baked in, so we don't need those in the sysroot.

The first iteration of this bbappend attempted to solve these
constraints by copying the modules from ${D}/${libdir}/grub/ to
${DEPLOYDIR} in do_deploy(), then removing ${D}${prefix}. This had the
unfortunate side effect of breaking the build in certain cases, such as
clean builds or reexecuting do_deploy() without the other steps of the
build.

Instead, remove the unwanted files in do_install(), and append the
required modules to GRUB_BUILDIN to create a standalone grub image
without any external modules at all.

Change-type: patch
Signed-off-by: Joseph Kogut <joseph@balena.io>