initrdscripts: always use by-uuid symlink looking for flasher rootfs
If the device with flasher rootfs is slow to bring up and rootfs is defined
as UUID=xxx the waiting loop in rootfs initrd script would assume UUIDs have
just been regenerated and wait for a by-state symlink instead. This only works
for the OS - flasher does not use the dynamically generated UUIDs
therefore we always want to use the by-uuid link for it.

Change-type: patch
Signed-off-by: Michal Toman <michalt@balena.io>