balena: dissolve healthcheck-image-load into healthcheck script
Previously we were loading the healthcheck image from it's own service
that was previously made `PartOf=` the balena.service.
This was done in an attempt to help get the system unstuck when the
image is removed (like through manual pruning), which would cause the
healthcheck to trigger the engine to reboot until the load service was
restarted by hand.

Further investigation found a race between the first execution of the
engine healthcheck script (which needs the image to be loaded) and the
loader service itself, which would lead to a similar state, requireing
manual intervention.

This change moves the loading into the healthcheck script itself,
allowing us to remove the loader script and service entirely, skipping
the whole service ordering issue.

Backport-to: next
Connects-to: #2122
Change-type: patch
Signed-off-by: Robert Günzler <robertg@balena.io>