modemmanager:u-blox-switch: Rework the u-blox modem switch to ECM mode
We need to make sure the modem is completely initialized before sending
the AT commands that do the switch to ECM mode. To achieve this we
change the systemd service dependency to depend on ModemManager.service
and then we determine the modem state by using mmcli and querying the
modem power-state property.

This procedure will only be done once. After the modem is placed in ECM
mode, it will stay in this mode for future reboots so there will be no
other delays in bringing up the modem connection.

Change-type: patch
Signed-off-by: Florin Sarbu <florin@balena.io>