Add oneshot service to migrate supervisor state config
The hostapp-update hooks would run before the supervisor update step
when using balenahup via the dashboard. This resulted in the balena
supervisor state conf file being out-of-date.

This commit removes the hostapp-update hook and instead uses a
oneshot service to ensure the conf is migrated if it does not exist
in the new path before starting the supervisor.

Change-type: patch
Signed-off-by: Kyle Harding <kyle@balena.io>