image-balena: Add check for docker image size
Currently it is possible to build a docker image that will then fail to
hup due to a lack of space on the inactive partition.

Add a task to check the docker image space against the available space on
the root filesystem partition and fail the build if it's too big.

Change-type: patch
Signed-off-by: Alex Gonzalez <alexg@balena.io>