kernel-headers-test: Update base image to buster
Since we don't have devices using older 3.x kernels we update to a newer
base image so that we don't have problems compiling this test kernel
module on newer kernels.
This avoids a compile error on kernel 5.10.31 on arm64
raspberrypicm4-ioboard for example where it would complain _mcount is
undefined because the older gcc in the intel-nuc-debian:stretch-20190717
base image would not create the _mcount symbol when compiling against
the newer 5.10.31 kernel.

Change-type: patch
Signed-off-by: Florin Sarbu <florin@balena.io>