linux-firmware: don't package iwlwifi below min api
balena-linux-firmware used a Python function to find and delete versions
of iwlwifi firmware below a given API version, in order to save space
when they're not required. Unfortunately, this causes problems with
pseudo aborting because the files have been removed outside of the
fakeroot context.

Instead of removing these files, set FILES_* to the appropriate paths of
the firmware blobs we'd like to keep.

At the same time, remove the IWLWIFI_FW_TOCLEAN variable. If a chipset
is present in IWLWIFI_FW_MIN_VERSION, that chipset's limited to
packaging blobs equal to or greater than the min version.

Change-type: patch
Signed-off-by: Joseph Kogut <joseph@balena.io>