recipes-connectivity: improve NTP dispatcher script
Change the NetworkManager NTP dispatcher script to update the
on/offline status of the NTP sources on 'connectivity-change' events
instead of 'up/down' events.

Currently 'chronyc onoffline' is run for 'up/down' events on any
network interface. It makes more sense to run it for
'connectivity-change' events as we are really interested in whether
the internet is there or not rather than whether an interface is
up or down.

Change-type: patch
Changelog-entry: recipes-connectivity: improve NTP dispatcher script
Signed-off-by: Mark Corbin <mark@balena.io>