resin-init-flasher: Use flasher kernel to emulate stage2 bootloader with LUKS
In order to avoid the need to unlock encrypted partitions in GRUB we want
to use a custom stage2 bootloader. Since that is not ready yet, emulate that
by copying flasher kernel to the EFI partition, start it from GRUB, have it
unlock all the partitions, find the real kernel and kexec into it.

Change-type: patch
Signed-off-by: Michal Toman <michalt@balena.io>