hostapp-update-hooks: Fix blacklisted extlinux.conf file path
The extlinux.conf file path in meta-balena is currently
incorrect, let's update it's boot partition path so that
BalenaHUP won't need to transfer the isolcpus setting to
the new OS anymore. This is necessary only when updating from older
OS versions in which the supervisor adds the isolcpus
setting to the boot partition extlinux.conf file.

Change-type: patch
Signed-off-by: Alexandru Costache <alexandru@balena.io>