resindataexpander: expand fs independent of partition
When resizing the filesystem fails, such as when resize2fs won't touch
it because it's dirty, the partition gets resized, but not the
filesystem. The script will not attempt to resize the filesystem again,
as it detects the partition has already been resized.

Split these actions apart, so that the filesystem resize is always
attempted. If resize2fs detects that the filesystem is already filling
available space, it will exit with no action taken.

Change-type: patch
Signed-off-by: Joseph Kogut <joseph@balena.io>