common: conf: don't disable user-ns by default
Upon reviewing the kernel configs for existing devices, we found at
least 64 DTs that previously enabled user namespacing, and only eight
that do not enable user namespacing.

In order to safely enable this config in the kernel while avoiding
enabling it for devices which previously did not have it, remove the
runtime configuration that disables it by default in meta-balena. We'll
update the few device repos that didn't enable CONFIG_USER_NS to disable
the feature at runtime.

Change-type: patch
Signed-off-by: Joseph Kogut <joseph@balena.io>