rollback-parse-bootloader: Revert to grub.cfg if grubenv not available
Since meta-balena 76c82dd987215982dbfcbdf950588dafdc4c129e the grub
environment variables have moved from grub.cfg to grubenv.

However, when a hostOS update fails and the system rolls back, these
grub environments are used to restore the contents of the boot partition
to the old OS which does not have a grubenv file.

This commit will fallback to looking for a grub.cfg file if grubenv is
not present in the filesystem

Fixes #2350

Change-type: patch
Signed-off-by: Alex Gonzalez <alexg@balena.io>