connectivity: proxy: move nadoo/glider to container
Previously, the core service exposed a /proxy endpoint that would start
up a proxy remotely, which would be used by a test in the connectivity
module. However, the endpoint returned the address for the testbot to be
used as the proxy in the response, and this required manual
configuration of the interface. Additionally, it requires the worker
service to install and provide glider for forward proxying.

Move the proxy (glider) to a container on the device being tested,
which simplifies configuration, and reduces the complexity and size of
the interface of Leviathan.

Change-type: patch
Signed-off-by: Joseph Kogut <joseph@balena.io>