connectivity: reduce ping interval to minimum
The interface test uses a simple ping to ensure a specific interface
works. It sends ten packets, and expects ten packets back. However, the
default interval is one second, which increases the time taken for the
test while not adding anything of value.

Reduce the timeout to the minimum non-privileged interval of 2ms.

Change-type: patch
Signed-off-by: Joseph Kogut <joseph@balena.io>