tests: wait for the chronyd service become active
Add a test to wait for the chronyd service to become active before
starting the time sync tests.

The current chronyd tests assume that the chronyd service is
instantly available at boot time. With the addition of the new
HTTPS time synchronisation service the starting of chronyd can be
delayed by a few seconds so we need to ensure that the service
is actually running before proceeding.

Change-type: patch
Changelog-entry: tests: wait for the chronyd service become active
Signed-off-by: Mark Corbin <mark@balena.io>