hostapp-update-hooks: Add a hook that aborts HUP to unsigned OS under secure boot
When a device is running in secure boot mode, it must not be possible to HUP
to an unsigned version of the OS because UEFI would refuse to boot it before
any of our self-recovering rollback mechanisms can be triggered. This would
effectively brick the device, needing physical access to recover.

Change-type: patch
Signed-off-by: Michal Toman <michalt@balena.io>