Use skopeo to push image to DUT registry
This simplifies the entire process at the cost of a new dependency in
the core image. We tell the DUT to HUP to an image stored in a localhost
registry, which obsoletes the need to secure that registry or tell the
DUT docker daemon to accept non-tls registry connections.

Change-type: patch
Signed-off-by: Robert Günzler <robertg@balena.io>