sign-efi.class, sign-kmod.class: Replace original files with signed ones
Currently the two classes would keep the original files untouched and store
the signed versions as .signed. This patch reverses the logic - the signed
files replace the original ones and the unsigned version is stored as .unsigned.
This is because there is no real use-case for the unsigned files, we always
want to ship the signed version, even if the particular DT does not require it,
this causes no harm.

Change-type: patch
Signed-off-by: Michal Toman <michalt@balena.io>