0-signed-update HUP hook: mount efivarfs if necessary
The hook tries to read EFI variables from efivarfs but this is not always
mounted within the container. We have already validated that we are running
in EFI mode therefore we can just check whether it is already mounted
and eventually mount with no further checks.

This also adds graceful handling of nonexistent variables since not all
UEFI implementations come with secure boot support.

Change-type: patch
Signed-off-by: Michal Toman <michalt@balena.io>