kernel-image-initramfs.bb: Ship kernel and matching signature
This patch replaces the kernel being shipped with the one that we eventually
sign for EFI - without signing the original file used would be identical
but after applying signature, the signed version is the one we want to ship.

It also fixes the file name for the detached signature, which must match
the name of the associated file.

Change-type: patch
Signed-off-by: Michal Toman <michalt@balena.io>