u-boot: Move u-boot configs inclusion into resin-u-boot.bbclass
Some BSPs name the u-boot such as u-boot-imx_2020 for example.
Because of that, our existing bbappend called u-boot_%.bbappend
will not match it and will fail to be applied.
This PR is a first pass at solving this by moving at least the
SRC_URI appending part from this bbappend and into the
resin-u-boot.bbclass instead. Further investigation is needed
to see if we can remove this bbappend altogether since we may
have all boards using Poky Warrior or newer, making this
u-boot bbappend obsolete.

Change-type: patch
Signed-off-by: Florin Sarbu <florin@balena.io>