initrdscripts: Use /run as bootparam_root storage
From v2.49, the hostapp-update utility creates the /run directory in the
root filesystem, however when huping from previous versions /run is not
there.

Commit bab3cd7f50022127bfef50fde9cd445b6b55a7b2 switches to use /tmp
to store the new UUID for the root partition on first boot after
generating new UUIDs as this is backwards compatible.

However, this means that the udev database in the initramfs is recreated
on the final system instead of reused. This becomes a problems for DM
devices (used in luks based disk encryption), as they are not re-processed
by udevd.

This change will use /run if available, so new releases that may implement
disk encryption work, or /tmp if not so it still remains backwards
compatible for HUP from older releases.

Change-type: patch
Signed-off-by: Alex Gonzalez <alexg@balena.io>