resin-mounts: Mount EFI partition under /mnt/efi instead of /mnt/boot/EFI
At this moment the EFI partition would be mounted under /mnt/boot/EFI
which results into duplication of the /EFI directory - what should be under
/mnt/boot/EFI is actually in /mnt/boot/EFI/EFI and this is a problem for HUP.

With this patch the EFI partition is mounted under /mnt/efi and /mnt/boot/EFI
is replaced by a symlink that points to the correct directory.

This only applies to DTs that have boot and EFI partitions split (e.g. LUKS)

Change-type: patch
Signed-off-by: Michal Toman <michalt@balena.io>