Add PCR protection policy to TPM operation unlocking LUKS passphrase
This patch adds the check for PCRs
- 0 - firmware/UEFI blob
- 1 - firmware/UEFI config
- 2 - Expansion card firmware blobs
- 3 - Expansion card firmware configs

The full protection is only applied on first boot after provisioning,
flasher only locks against PCRs 0, 2 and 3.
This is because when using flasher, the environment is not in the same
state as the one we want to lock to. In particular the boot order is
different (flasher is booted from a different drive than the resulting OS).
As for now we were not able to find a better solution than fully locking
only on first boot. This means the device must be booted in a secure
environment at least once after provisioning.

Change-type: patch
Signed-off-by: Michal Toman <michalt@balena.io>