u-boot: Move config fragments merging code out of common layer
Let's only apply this code to pre-warrior as intended. Starting
with warrior this code is in poky so we can rely on it from
there instead of keeping a duplicate in meta-balena-common.
This helps with BSPs that may define their own tasks (configure,
compile and so on) for u-boot (keeping this fragments merging
code in meta-balena-common would have the effect of overwriting
these tasks from here and thus breaking the build).

Change-type: patch
Signed-off-by: Florin Sarbu <florin@balena.io>