Remove config-json.target
Unit files that are configured via config.json and want to be restarted
on config.json changes are part of this target.

This causes a burst of service restarts each time config.json changes,
independently of whether the configuration changes applies to the unit
or not.

This commit removes all config-json.target instances in preparation for
the introduction of a better, more fine grained mechanism.

Change-type: minor
Signed-off-by: Alex Gonzalez <alexg@balena.io>