suites/os: Add under-voltage test
We add this this check since we discovered
that with newer boot firmware the Pi4 will
not reboot when under-voltage is reported.

We leave this check to be performed on all
devices to avoid missing any RPi derivate
devices that may be added in the future,
especilly since it's just a kernel log
check.

We also don't mark the test as failed in
this case, because there may be devices that
continue to work in the under-powered state and we
don't want to block or delay development if the DUT
is remote and cannot be checked or replaced easily.

This check provides an indication when particular tests
fail, like for instance if the device doesn't come up
after a reboot or a hostOS update.

Signed-off-by: Alexandru Costache <alexandru@balena.io>