suites/hup: Add under-voltage test before and after HUP
We add this this check since we discovered
that with newer boot firmware the Pi4 will
not reboot when under-voltage is reported.

We don't mark the test as failed in
this case, since there may be devices that
continue to work even if under-powered and we
don't want to block or delay development if the
DUT is remote and cannot be checked or replaced
easily.

This check provides an indication for further
debugging of failed test cases.

Change-type: patch
Signed-off-by: Alexandru Costache <alexandru@balena.io>