openvpn: Remove dependency on timesync-http target
We want the VPN to start unconditionally even if the connectivity URL
is not reachable.

The rationale for adding the dependency in the first place was that
without an initial timesync certificate checks may fail. This can still
happen, but the VPN will retry continuously and eventually succeed once
the time is synched.

What happens now is that the VPN is delayed until the http sync, so if
the connectivity URL is blocked or unreachable, but the internet is
accessible, VPN will not even launch and no remote debugging is possible.

Fixes #2508

Change-type: patch
Signed-off-by: Alex Gonzalez <alexg@balena.io>