suites/os: Add testcase for RPi device-tree
Upstream kernel added a new fkms overlay for
the Pi4 and CM4 and switched to using it
from overlay-map, however, we were not aware
of this.

Let's add a vcdbg testcase to warn in case the RPi
firmware fails to load any overlay or dtb, and set
gpu_mem to 64 so that vcdbg can run.

Change-type: patch
Signed-off-by: Alexandru Costache <alexandru@balena.io>