prepare-openvpn: Make configurable
The prepare-openvpn service needs to be restarted to regenerate the
VPN keys when provisioning is finished and the `apiKey` is removed.

Note that flasher images make changes to `/mnt/boottmp/config.json` so
we still need to manually kick `prepare-openvpn` in
`resin-device-register`, which is preferable to modifying services to
watch multiple paths.

Change-type: patch
Signed-off-by: Alex Gonzalez <alexg@balena.io>