grub-efi: Remove patch passing secure boot status to the kernel
This was originally introduced in combination with a kernel patch backported
from Red Hat kernel that would enable kernel lockdown when secure boot
is enabled. We have since changed the approach, dropped the kernel patch
and when in secure boot mode use a different GRUB config file that enables
lockdown on kernel command line unconditionally.

That said, while the patch works fine, we do not really need it and it adds
extra overhead porting it to newer yocto versions so there is no point keeping
it at this moment. If in the future there is need for the kernel to know
whether it is in secure boot mode or not, we can roll this back and rebase.

Change-type: patch
Signed-off-by: Michal Toman <michalt@balena.io>