Disable Engine startup timeouts
There are known situations in which balenaEngine times out during
initialization (for example, during aufs to overlayfs migrations, or
when restarting a device that was running a large number of containers).
When these time outs occur, Systemd kills the Engine, causing further
problems.

To avoid these cases, this commit disables timeouts during the Engine
initialization.

This is also aligned with the default Systemd settings distributed with
the Moby project.

Signed-off-by: Leandro Motta Barros <leandro@balena.io>
Change-type: patch