networkmanager: Use default DHCP timeout
This reverts 5047757 where we set the global DHCP timeout for ipv4 to
infinity to mitigate problems with routers that do not respond to DHCP
requests in time.

This however causes issues on some routers, which after power cycle fail to
renew leases after they are expired.

The same ipv4.dhcp-timeout can be specified on per connection basis. Since
the case with lease renewal due to router power cycle is a more common case,
it is better to keep the global dhcp-timeout default unmodified. If the
setting needs adjustment then it can be overwritten in the particular
connection profile.

Change-type: patch
Signed-off-by: Zahari Petkov <zahari@balena.io>