tests: os: config-json: fix race by waiting for InvocationID change
When tests execute quicker, a race condition can occur where config.json
is edited to remove the dnsServers property, but /run/dnsmasq.servers is
not changed yet. This causes the test to fail, as the file is not empty.

    not ok 1 - We should have an empty /run/dnsmasq.servers file.
      ---
      found: |-
	server=1.1.1.1
	server=1.1.1.1
      pattern: !re /^\s?$/
      diff: |
	--- expected
	+++ actual
	-/^\s?$/
	+"server=1.1.1.1\nserver=1.1.1.1"

Fix this by waiting until the InvocationID of dnsmasq.service changes.

Change-type: patch
Signed-off-by: Joseph Kogut <joseph@balena.io>