balena-efi.service: Mount if /mnt/boot/EFI is a symlink
This changes the condition in the unit file from checking whether
/dev/disk/by-state/balena-efi exists to checking whether /mnt/boot/EFI
is a symlink. The original approach has a race condition populating
the by-state symlink - it is depending on udev and if the link is
not present when the service is started (after the boot partition is mounted),
the service fails and the EFI partition is never mounted.

The new approach does the trick pretty well - /mnt/boot/EFI is a symlink
if the EFI partition is split and a regular directory in case there is a single
boot partition. That said the service is only started when necessary
and the waiting for udev is implemented as a part of the mount script.

Change-type: patch
Signed-off-by: Michal Toman <michalt@balena.io>