chronyd: Add periodic time synchronization check
By default, chronyd is configured to poll every ~4h to minimize bandwidth
use. However, if the sync fails, it won't retry until the next poll time.

This means that time can get out of sync on unreliable networks.

This commit introduces a timer that checks chrony synchronization status
on a varying time range between 1m and 15m, roughly equivalent to the
chronyd minpoll and maxpoll defaults which are 64s and 1024s respectively.
It checks synchronization once every 10s for 10 minutes, and syncs time
if not synchronized.

Fixes #2314

Change-type: minor
Signed-off-by: Alex Gonzalez <alexg@balena.io>