tests: os: make apiKey an optional parameter
When accessing a test device as part of a fleet, a cloud API key is
required in order to generate an SSH key and access the device through
the VPN. However, when accessing a device locally, such as a QEMU
instance running on the workstation itself, we have a direct path, and
no VPN is necessary.

Make the apiKey optional, and don't login when it's not specified. This
allows direct connections to work without it.

Change-type: patch
Signed-off-by: Joseph Kogut <joseph@balena.io>