Use a lightweight Engine healthcheck
Previously, our healthcheck verified if we were able to start a new
container. This had two downsides:

1. It was relatively heavyweight. In devices under heavy load, it would
   sometimes take so long to run that the Engine was killed by the
   watchdog.
2. It wrote to the storage media. Creating a container involves writing
   some data to persistent storage, therefore the healthcheck was
   wearing the storage media.

This new healthcheck simply pings both `balenad` and `containerd`, which
is much faster than starting a new container and doesn't write to disk.
The step of pinging `containerd` is important because we have seen at
least one case in the past in which `balenad` was working but
`containerd` wasn't.

Signed-off-by: Leandro Motta Barros <leandro@balena.io>
Change-type: patch