tests: cloud: supervisor: use minimal app image
Several supervisor tests in the managed cloud suite depend on the
balena-nodejs-hello-world app, which builds a ~225 MB image for testing.
Building, pushing, and pulling this image is time consuming. Switch to a
minimal image that accomplishes much the same thing.

This cuts approximately a minute from the supervisor tests.

Change-type: patch
Signed-off-by: Joseph Kogut <joseph@balena.io>