tests: cloud: multicontainer: use minimal app image
The multicontainer tests clone and build the application from
balena-io-examples/multicontainer-getting-started, which is ~420 MB
total. Reuse the same images from the supervisor tests, which are just
upstream alpine images with an entry script.

This shaves off ~30s from the multicontainer tests, possibly more or
less depending on upload speeds and container build times.

Signed-off-by: Joseph Kogut <joseph@balena.io>