meta-balena-rust: Add rust recipes to keep the rust version a distribution config
With Kirkstone and the support of openSSL 3.0 it's not possible to find
a set of dependencies that work for all of our rust applications across
the 1.32 to 1.62 toolchain versions that are supported across all the
Yocto versions we keep compatibility with.

This layer allows to set a preferred version as a distro setting that can
be used across all Yocto versions.

These rust recipes come from Poky Kirkstone
2642bdb8908afb800e6dc5722602a87228dd8694

Change-type: patch
Signed-off-by: Alex Gonzalez <alexg@balena.io>