meta-balena-rust: Modify to use without oe.rust.arch_to_rust_arch() support
The arch_to_rust_arch() function was introduced to cater for ppc64le
systems and is not present in older Yocto releases.

This commit reverts to the previous way of setting the rust architecture.
It will not work for ppc64le and if we would need to support such an
architecture the arch_to_rust_arch() function will still be called if it
exists in Kirkstone or newer Yocto versions.

Change-type: patch
Signed-off-by: Alex Gonzalez <alexg@balena.io>