balena-config-vars: Set permissions for cache file
The sshd daemon is configured to fetch keys from the API for local
user connections. The script that fetches the keys, cloud-public-sshkeys,
sources balena-config-vars and is run as an exclusive non-root user.

Let's set the correct permissions for this file to allow not to break
the above.

Also, add a test to check that the SSH authentication with user
registered keys does not break again.

Fixes #2785

Change-type: patch
Signed-off-by: Alex Gonzalez <alexg@balena.io>