classes: kernel-balena: improve aufs branch selection
The task that applies the aufs patches is still required to support
hostOS updates between aufs and overlay2 balenaOS versions.

This commit adds support for 5.15 kernels and improves the branch
selection logic to cover some corner cases.

It also adds a test task so that we can test regressions.

Change-type: patch
Signed-off-by: Alex Gonzalez <alexg@balena.io>