initrdscripts: add recovery module
When `recovery` is passed in the kernel command line, the initramfs will
not boot into the OS but launch adbd.

A connection can then be established via the network (if a DHCP address
is provided) or using a USB gadget connection.

For a network connection, the client needs to install `adb` and run:
```
adb connect $IPADDR:5555
adb shell
```

Other adb commands, like pull/push of files are also available.

Resolves #2864

Change-type: minor
Signed-off-by: Alex Gonzalez <alexg@balena.io>