copy btrfs-tools from kirkstone
Yocto parses the original recipe before the bbappend is applied, which
results in the recipe still inheriting from distutils3-base if the
python package is removed in an append. See the mailing list thread [0]
for details.

Instead, copy the recipe into our layer unmodified so we can remove
unneeded packages. This keeps the initramfs from blowing up in size.

[0] https://lore.kernel.org/all/aebe57d858d895a9266a9a25e90b60c8a34dd9e3.camel@redrectangle.org/T/

Change-type: patch
Signed-off-by: Joseph Kogut <joseph@balena.io>