common: kernel-devsrc: fix pseudo abort
Yocto builds will abort due to inconsistencies between the pseudo
database and files modified outside of the pseudo context [0].

This will occasionally cause builds to fail in the do_deploy step of the
kernel-devsrc recipe. [1]

Fix this by not removing the kernel_source tarball in the do_deploy
step.

Fixes: https://github.com/balena-os/meta-balena/issues/2806

[0] https://wiki.yoctoproject.org/wiki/Pseudo_Abort
[1] https://github.com/balena-os/meta-balena/issues/2806

Change-type: patch
Signed-off-by: Joseph Kogut <joseph@balena.io>