test: ssh-auth: fix test cases using custom keys
There are two sets of keys used in this test, one stored in `/root/id`
which is created by the cloud suite to SSH via the proxy server, and
a custom key stored in `/root/test_id` used in some of the subtests.

Fix the test cases using the custom key to use the correct private key.

Change-type: patch
Signed-off-by: Alex Gonzalez <alexg@balena.io>