ssh-auth: place custom key in /tmp folder
The connection between worker and DUT is different when running on qemu
or testbot. On QEMU there is a direct connection, while on testbot there
is a tunnel to specific ports established, and SSH keys are copied to
the `/tmp` folder in the worker.

Making the custom key live by default in the `/tmp` folder means that
we can use a common location for both cases.

Change-type: patch
Signed-off-by: Alex Gonzalez <alexg@balena.io>