ssh-auth: do not use a separate custom key
Given that testbot devices use a tunnel to specific ports to communicate
with the DUT that is established with the suite-generated keys, using
a different keypair for the ssh-auth test would require to tear down and
re-establish the tunnel.

It's easier to just use the existing key pair in the ssh-auth test.

Change-type: patch
Signed-off-by: Alex Gonzalez <alexg@balena.io>