networkmanager: Update to 1.40.4
Main NetworkManager recipe was migrated to match latest one from
meta-openembedded, so that all improvements are merged now.

Excluded from it are `iwd` and `dhcpcd` daemon configurations that are
not used by us.

Default NM firewall in meta-openembedded is `nftables` where we are still
using `iptables`.

The new recipe relies on `meson` as a build system now.

The .bbapend file that contains modifications specific to balena is preserved.
Only `balena-client-id.patch` is removed as it references code that no longer
exists. This is because the internal systemd DHCPv4 client code that NM used
is now replaced by nettools' n-dhcp4 implementation.

Closes #2872

Change-type: minor
Changelog-entry: Update NetworkManager to 1.40.4
Signed-off-by: Zahari Petkov <zahari@balena.io>