resin-init-flasher: replace shutdonw for reboot in case of migration
When migrating, we can't assume a device can be remotely powered on, so
finish the flashing with a reboot instead of a shutdown.

This is possible as the initramfs does not currently have a shutdown
command and avoids the complexity of having to detect whether we are
running from initramfs or not.

Change-type: patch
Signed-off-by: Alex Gonzalez <alexg@balena.io>